app/requirements.txt
psutil
paho-mqtt
nltk